cloud/functions/requirements.txt
Package           Version
----------------- -------
annotated-types   0.7.0
azure-functions   1.21.3
greenlet          3.1.1
pip               23.0.1
pydantic          2.10.6
pydantic_core     2.27.2
setuptools        65.5.0
SQLAlchemy        2.0.38
sqlmodel          0.0.23
typing_extensions 4.12.2
